+++ src/gui/helpers.py
@@ -270,15 +270,16 @@ def custom_save(sender, app_data, user_data) -> None:
     try:
 
         custom_save_file.addMany(
-            [{
-                "save_name": dpg.get_value(item="save_custom_input"),
-                "channel": channel,
-                "Power": dpg.get_value(f"power_{channel}"),
-                "Bandwidth": dpg.get_value(f"bandwidth_{channel}"),
-                "Frequency": dpg.get_value(f"freq_{channel}"),
-                "Date": dt_string,
-            }]
-            for channel in range(1, 9)
+            [
+                {
+                    "save_name": dpg.get_value(item="save_custom_input"),
+                    "Power": dpg.get_value(f"power_{channel}"),
+                    "Bandwidth": dpg.get_value(f"bandwidth_{channel}"),
+                    "Frequency": dpg.get_value(f"freq_{channel}"),
+                    "Date": dt_string
+                }
+                for channel in range(1, 9)
+            ]
         )
 
     except (
@@ -287,6 +288,7 @@ def custom_save(sender, app_data, user_data) -> None:
         KeyError,
         errors.db_errors.SchemaError,
         AttributeError,
+        json.JSONDecodeError,
     ):
         loggey.exception(msg=f"database failure | {custom_save.__name__}()")
 
@@ -305,18 +307,31 @@ def custom_load() -> list:
     except FileNotFoundError:
         loggey.exception(msg="No custom save file found")
 
-    return custom_save_file.getAll()
+    try:
+        return custom_save_file.getAll()
+    except json.JSONDecodeError:
+        loggey.exception(msg="No save file detected OR save file corrupted")
+        loggey.warning(
+            msg="Save file being overwritten due to corruption or nonexistence")
+        with open(file="db/long_save.json", mode="w") as overwrite_corruption_or_create:
+            overwrite_corruption_or_create.write('{"data":[]}')
+        return custom_save_file.getAll()
 
 
 def load_chosen(sender=None, app_data=None, user_data=None) -> None:
     """Take in the chosen file to be loaded"""
 
-    print(user_data)
     _custom_load = db.getDb("db/long_save.json")
     _ret_data: list[dict[str]] = _custom_load.getBy({"save_name": user_data})
 
-    for data in _ret_data:
-        print(data)
+    [
+        (
+            dpg.set_value(item=f"freq_{iter}", value=data["Frequency"]),
+            dpg.set_value(item=f"power_{iter}", value=data["Power"]),
+            dpg.set_value(item=f"bandwidth_{iter}", value=data["Bandwidth"])
+        )
+        for iter, data in enumerate(_ret_data, start=1)
+    ]
 
 
 def auto_fill_freq(