+++ CHANGELOG.md
@@ -16,9 +16,12 @@ Here we write upgrade and change notes.
 
 ✳️ PATCH version when you make backwards compatible bug fixes.
 
+
+
 ## ✳️[0.12.1] - 2022 JUN 7
 
 - Critical error in `find_dev.sh`; linux device finding listing script
+- Added YAML file to run PyPi update using GitHub Actions 
 
 ## ✴️[0.12.0] - 2022 JUN 6