+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Here we write upgrade and change notes.
 
 ## ✳️[0.12.1] - 2022 JUN 7
 
+- Removed the `find_dev.sh` script completely and retained that functionality
 - Critical error in `find_dev.sh`; linux device finding listing script
 
 ## ✴️[0.12.0] - 2022 JUN 6