+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Here we write upgrade and change notes.
 
 - Removed the `find_dev.sh` script completely and retained that functionality
 - Critical error in `find_dev.sh`; linux device finding listing script
+- All phonetic mission buttons are configurable via config files
 
 ## ✴️[0.12.0] - 2022 JUN 6